,cleaning code of the stack
