+++ lib/helpertools/README.md
@@ -1,0 +1,18 @@
+# HelperTools
+
+A collection of helper functions and classes for the lambda functions 
+
+## Installation
+In a shell, from the folder of this README.md, install the module with
+```bash
+cd .. && pip install helpertools
+```
+If you want to force a full reinstall, use instead
+```bash
+cd .. && pip install --force-reinstall helpertools
+```
+
+## Usage
+```python
+from helpertools import get_logger #and other functions
+```
