+++ lib/textracttools/setup.py
@@ -13,6 +13,7 @@ setuptools.setup(
     long_description_content_type="text/markdown",
     # url="https://github.com/pypa/sampleproject",
     packages=setuptools.find_packages(),
+    install_requires=['wheel'],
     classifiers=[
         "Programming Language :: Python :: 3",
     #     "License :: OSI Approved :: MIT License",
