+++ selectable_pdf_infra/app.py
@@ -9,7 +9,7 @@ from selectable_pdf_infra.selectable_pdf_stack import SelectablePdfStack
 stack_id = 'SelectablePDF'
 stack_desc = 'infrastructure to generate selectable PDF at scale'
 # Region
-region = 'eu-west-1'  #Irland
+region = 'eu-west-1'
 # logging level for the lambda functions
 log_level = 'INFO'  #log level for the Lambdas. only INFO is implemented atm.
 
@@ -22,7 +22,7 @@ infra_stack = SelectablePdfStack(
     construct_id=stack_id, 
     log_level=log_level,
     description=stack_desc,
-    env={"region": region},
+    env={'region': region},
 )
 
 app.synth()
