+++ selectable_pdf_infra/lambda/start_textract/main.py
@@ -24,26 +24,14 @@ import random
 import string
 from datetime import datetime, timedelta
 from urllib.parse import unquote_plus
+
+from helpertools import get_logger
+
 from typing import Dict
 
-# prepare the logger. If no LOG_LEVEL env var or wrong LOG_LEVEL env var, fallback 
+#If no LOG_LEVEL env var or wrong LOG_LEVEL env var, fallback 
 # to INFO log level
-log_level = os.getenv('LOG_LEVEL', default='INFO')
-log_level_int = int()
-if log_level=='WARNING':
-    log_level_int = logging.WARNING
-elif log_level=='ERROR':
-    log_level_int = logging.WARNING
-elif log_level=='DEBUG':
-    log_level_int = logging.DEBUG
-else:
-    log_level_int = logging.INFO
-logging.basicConfig(
-    format='%(levelname)s %(message)s',
-    level=log_level_int,
-    force=True  #new in py3.8
-)
-logger = logging.getLogger()
+logger = get_logger(os.getenv('LOG_LEVEL', default='INFO'))
 
 
 def lambda_handler(event, context):
