+++ selectable_pdf_infra/lambda_layer/helpertools_py38/README.md
@@ -1,6 +1,6 @@
-# textracttools_p38
+# helpertools_p38
 
-Build an AWS Lambda Layer with the TextractTools module. To create the layer, 
+Build an AWS Lambda Layer with the HelperTools module. To create the layer, 
 update the module source in `createlayer.sh` (if required), then run:
 ```bash
 ./createlayer.sh 3.8
