--- selectable_pdf_infra/lambda_layer/pypi_py38/requirements.txt
@@ -1,2 +1,0 @@
-
-PyMuPDF
